modules/chat: cpu optimizations
